Decouple config from admin
