+++ projects/plugins/boost/app/admin/class-admin.php
@@ -27,21 +27,9 @@ class Admin {
 	const MENU_SLUG = 'jetpack-boost';
 
 	/**
-	 * Nonce action for setting the statuses of rating and score prompts.
-	 */
-	const SET_SHOW_RATING_PROMPT_NONCE = 'set_show_rating_prompt';
-	const SET_SHOW_SCORE_PROMPT_NONCE  = 'set_show_score_prompt';
-
-	/**
 	 * Option to store options that have been dismissed.
 	 */
 	const DISMISSED_NOTICE_OPTION = 'jb-dismissed-notices';
-
-	/**
-	 * Name of option to store status of show/hide rating and score prompts
-	 */
-	const SHOW_RATING_PROMPT_OPTION = 'jb_show_rating_prompt';
-	const SHOW_SCORE_PROMPT_OPTION  = 'jb_show_score_prompt';
 
 	/**
 	 * Main plugin instance.
@@ -57,17 +45,26 @@ class Admin {
 	 */
 	private $speed_score;
 
+	/**
+	 * Configuration constants.
+	 *
+	 * @param Config $config
+	 */
+	private $config;
+
+
 	public function __construct( Optimizations $modules ) {
 		$this->modules     = $modules;
 		$this->speed_score = new Speed_Score( $modules );
 		Environment_Change_Detector::init();
 		Premium_Pricing::init();
 
+		$this->config = new Config;
+		$this->config->init();
+
 		add_action( 'init', array( new Analytics(), 'init' ) );
 		add_filter( 'plugin_action_links_' . JETPACK_BOOST_PLUGIN_BASE, array( $this, 'plugin_page_settings_link' ) );
 		add_action( 'admin_notices', array( $this, 'show_notices' ) );
-		add_action( 'wp_ajax_set_show_rating_prompt', array( $this, 'handle_set_show_rating_prompt' ) );
-		add_action( 'wp_ajax_set_show_score_prompt', array( $this, 'handle_set_show_score_prompt' ) );
 		add_filter( 'jetpack_boost_js_constants', array( $this, 'add_js_constants' ) );
 
 		$this->handle_get_parameters();
@@ -127,45 +124,10 @@ class Admin {
 			true
 		);
 
-		$optimizations = ( new Optimizations() )->get_status();
-		// Prepare configuration constants for JavaScript.
-		$constants = array(
-			'version'             => JETPACK_BOOST_VERSION,
-			'api'                 => array(
-				'namespace' => JETPACK_BOOST_REST_NAMESPACE,
-				'prefix'    => JETPACK_BOOST_REST_PREFIX,
-			),
-			'optimizations'       => $optimizations,
-			'locale'              => get_locale(),
-			'site'                => array(
-				'domain'    => ( new Status() )->get_site_suffix(),
-				'url'       => get_home_url(),
-				'online'    => ! ( new Status() )->is_offline_mode(),
-				'assetPath' => plugins_url( $internal_path, JETPACK_BOOST_PATH ),
-			),
-			'shownAdminNoticeIds' => $this->get_shown_admin_notice_ids(),
-			'preferences'         => array(
-				'showRatingPrompt' => $this->get_show_rating_prompt(),
-				'showScorePrompt'  => $this->get_show_score_prompt(),
-				'prioritySupport'  => Premium_Features::has_feature( Premium_Features::PRIORITY_SUPPORT ),
-			),
-
-			/**
-			 * A bit of necessary magic,
-			 * Explained more in the Nonce class.
-			 *
-			 * Nonces are automatically generated when registering routes.
-			 */
-			'nonces'              => Nonce::get_generated_nonces(),
-		);
-
-		// Give each module an opportunity to define extra constants.
-		$constants = apply_filters( 'jetpack_boost_js_constants', $constants );
-
 		wp_localize_script(
 			$admin_js_handle,
 			'Jetpack_Boost',
-			$constants
+			$this->config->constants()
 		);
 
 		wp_set_script_translations( $admin_js_handle, 'jetpack-boost' );
@@ -203,12 +165,17 @@ class Admin {
 	}
 
 	/**
-	 * Check for permissions.
-	 *
-	 * @return bool
+	 * Get the list of dismissed notices.
 	 */
-	public function check_for_permissions() {
-		return current_user_can( 'manage_options' );
+	public static function get_dismissed_notices() {
+		return \get_option( self::DISMISSED_NOTICE_OPTION, array() );
+	}
+
+	/**
+	 * Delete the option tracking which admin notices have been dismissed during deactivation.
+	 */
+	public static function clear_dismissed_notices() {
+		\delete_option( self::DISMISSED_NOTICE_OPTION );
 	}
 
 	/**
@@ -221,7 +188,7 @@ class Admin {
 		$notices          = $this->get_admin_notices();
 
 		// Filter out any that have been dismissed, unless newer than the dismissal.
-		$dismissed_notices = \get_option( self::DISMISSED_NOTICE_OPTION, array() );
+		$dismissed_notices = self::get_dismissed_notices();
 		$notices           = array_filter(
 			$notices,
 			function ( $notice ) use ( $dismissed_notices ) {
@@ -243,27 +210,11 @@ class Admin {
 	}
 
 	/**
-	 * Returns an array of notice ids (i.e.: jetpack-boost-notice-[slug]) for all
-	 * visible admin notices.
-	 *
-	 * @return array List of notice ids.
-	 */
-	private function get_shown_admin_notice_ids() {
-		$notices = $this->get_admin_notices();
-		$ids     = array();
-		foreach ( $notices as $notice ) {
-			$ids[] = $notice->get_id();
-		}
-
-		return $ids;
-	}
-
-	/**
 	 * Returns a list of admin notices to show. Asks each module to provide admin notices the user needs to see.
 	 *
 	 * @return \Automattic\Jetpack_Boost\Admin\Admin_Notice[]
 	 */
-	public function get_admin_notices() {
+	public static function get_admin_notices() {
 		return apply_filters( 'jetpack_boost_admin_notices', array() );
 	}
 
@@ -279,7 +230,7 @@ class Admin {
 		if ( is_admin() && ! empty( $_GET['jb-dismiss-notice'] ) ) {
 			$slug = sanitize_title( wp_unslash( $_GET['jb-dismiss-notice'] ) );
 
-			$dismissed_notices = \get_option( self::DISMISSED_NOTICE_OPTION, array() );
+			$dismissed_notices = self::get_dismissed_notices();
 
 			if ( ! in_array( $slug, $dismissed_notices, true ) ) {
 				$dismissed_notices[] = $slug;
@@ -291,94 +242,12 @@ class Admin {
 	// phpcs:enable WordPress.Security.NonceVerification.Recommended
 
 	/**
-	 * Handle the ajax request to set show-rating-prompt status.
-	 */
-	public function handle_set_show_rating_prompt() {
-		if ( check_ajax_referer( self::SET_SHOW_RATING_PROMPT_NONCE, 'nonce' ) && $this->check_for_permissions() ) {
-			$response = array(
-				'status' => 'ok',
-			);
-
-			$is_enabled = isset( $_POST['value'] ) && 'true' === $_POST['value'] ? '1' : '0';
-			\update_option( self::SHOW_RATING_PROMPT_OPTION, $is_enabled );
-
-			wp_send_json( $response );
-		} else {
-			$error = new \WP_Error( 'authorization', __( 'You do not have permission to take this action.', 'jetpack-boost' ) );
-			wp_send_json_error( $error, 403 );
-		}
-	}
-
-	/**
-	 * Get the value of show_rating_prompt.
-	 *
-	 * This determines if there should be a prompt after speed score improvements. Initially the value is set to true by
-	 * default. Once the user clicks on the rating button, it is switched to false.
-	 *
-	 * @return bool
-	 */
-	public function get_show_rating_prompt() {
-		return \get_option( self::SHOW_RATING_PROMPT_OPTION, '1' ) === '1';
-	}
-
-	/**
-	 * Handle the ajax request to set show-rating-prompt status.
-	 */
-	public function handle_set_show_score_prompt() {
-		if ( check_ajax_referer( self::SET_SHOW_SCORE_PROMPT_NONCE, 'nonce' ) && $this->check_for_permissions() ) {
-			$response = array(
-				'status' => 'ok',
-			);
-
-			$is_enabled = isset( $_POST['value'] ) && 'true' === $_POST['value'] ? '1' : '0';
-			\update_option( self::SHOW_SCORE_PROMPT_OPTION, $is_enabled );
-
-			wp_send_json( $response );
-		} else {
-			$error = new \WP_Error( 'authorization', __( 'You do not have permission to take this action.', 'jetpack-boost' ) );
-			wp_send_json_error( $error, 403 );
-		}
-	}
-
-	/**
-	 * Get the value of show_score_prompt.
-	 *
-	 * This determines if there should be a prompt after speed score worsens. Initially the value is set to true by
-	 * default. Once the user clicks on the support button, it is switched to false.
-	 *
-	 * @return bool
-	 */
-	public function get_show_score_prompt() {
-		return \get_option( self::SHOW_SCORE_PROMPT_OPTION, '1' ) === '1';
-	}
-
-	/**
-	 * Delete the option tracking which admin notices have been dismissed during deactivation.
-	 */
-	public static function clear_dismissed_notices() {
-		\delete_option( self::DISMISSED_NOTICE_OPTION );
-	}
-
-	/**
-	 * Clear the status of show_rating_prompt
-	 */
-	public static function clear_show_rating_prompt() {
-		\delete_option( self::SHOW_RATING_PROMPT_OPTION );
-	}
-
-	/**
-	 * Clear the status of show_score_prompt
-	 */
-	public static function clear_show_score_prompt() {
-		\delete_option( self::SHOW_SCORE_PROMPT_OPTION );
-	}
-	/**
 	 * Clear a specific admin notice.
 	 *
 	 * @param string $notice_slug The notice slug.
 	 */
 	public static function clear_dismissed_notice( $notice_slug ) {
-		$dismissed_notices = \get_option( self::DISMISSED_NOTICE_OPTION, array() );
+		$dismissed_notices = self::get_dismissed_notices();
 
 		if ( in_array( $notice_slug, $dismissed_notices, true ) ) {
 			array_splice( $dismissed_notices, array_search( $notice_slug, $dismissed_notices, true ), 1 );
@@ -396,8 +265,8 @@ class Admin {
 	 */
 	public function add_js_constants( $constants ) {
 		// Information about the current status of Critical CSS / generation.
-		$constants['showRatingPromptNonce'] = wp_create_nonce( self::SET_SHOW_RATING_PROMPT_NONCE );
-		$constants['showScorePromptNonce']  = wp_create_nonce( self::SET_SHOW_SCORE_PROMPT_NONCE );
+		$constants['showRatingPromptNonce'] = wp_create_nonce( Config::SET_SHOW_RATING_PROMPT_NONCE );
+		$constants['showScorePromptNonce']  = wp_create_nonce( Config::SET_SHOW_SCORE_PROMPT_NONCE );
 
 		return $constants;
 	}
